Moved windows service to separate file
